+++ src/nidb/survey.h
@@ -48,7 +48,7 @@ public:
     QString notes;
     QString rater;
     QString visit;
-    qint8 status = -1;
+    qint16 status = -1;
     int instrumentRowID = -1;
     qint64 surveyRowID = -1;
